Refactored exception package and added FileSelectorException
