+++ src/main/java/tudelft/ti2806/pl3/exception/FileSelectorException.java
@@ -1,0 +1,7 @@
+package tudelft.ti2806.pl3.exception;
+
+/**
+ * Created by Boris Mattijssen on 26-05-15.
+ */
+public class FileSelectorException extends Throwable {
+}
